+++ redis/tests/test_basic.rs
@@ -157,7 +157,7 @@ fn test_info() {
     let info: redis::InfoDict = redis::cmd("INFO").query(&mut con).unwrap();
     assert_eq!(
         info.find(&"role"),
-        Some(&redis::Value::Status("master".to_string()))
+        Some(&redis::Value::SimpleString("master".to_string()))
     );
     assert_eq!(info.get("role"), Some("master".to_string()));
     assert_eq!(info.get("loading"), Some(false));